pyproject.toml
[project]
name = "weather-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-limiter>=4.0.0",
    "flask-smorest>=0.46.2",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
]
